+++ src/renderer/libs/exportRows.ts
@@ -1,14 +1,16 @@
 import { ClientCode } from 'common/interfaces/antares';
 import { jsonToSqlInsert } from 'common/libs/sqlUtils';
+import * as json2php from 'json2php';
 
 export const exportRows = (args: {
-   type: 'csv' | 'json'| 'sql';
+   type: 'csv' | 'json'| 'sql' | 'php';
    content: object[];
    table: string;
    client?: ClientCode;
    fields?: {
       [key: string]: {type: string; datePrecision: number};
    };
+   sqlOptions?: {sqlInsertAfter: number; sqlInsertDivider: 'bytes' | 'rows'; targetTable: string};
 }) => {
    let mime;
    let content;
@@ -21,27 +23,37 @@ export const exportRows = (args: {
          if (args.content.length)
             csv.push(Object.keys(args.content[0]).join(';'));
 
-         for (const row of args.content)
-            csv.push(Object.values(row).map(col => typeof col === 'string' ? `"${col}"` : col).join(';'));
+         for (const row of args.content) {
+            csv.push(Object.values(row).map(col => {
+               if (typeof col === 'string') return `"${col}"`;
+               if (col instanceof Buffer) return col.toString('base64');
+               if (col instanceof Uint8Array) return Buffer.from(col).toString('base64');
+               return col;
+            }).join(';'));
+         }
 
          content = csv.join('\n');
          break;
       }
       case 'sql': {
          mime = 'text/sql';
-         const sql = [];
+         const sql = jsonToSqlInsert({
+            json: args.content,
+            client:
+            args.client,
+            fields: args.fields,
+            table: args.sqlOptions.targetTable || args.table,
+            options: args.sqlOptions
+         });
 
-         for (const row of args.content) {
-            sql.push(jsonToSqlInsert({
-               json: row,
-               client:
-               args.client,
-               fields: args.fields,
-               table: args.table
-            }));
-         }
-
-         content = sql.join('\n');
+         content = sql;
+         break;
+      }
+      case 'php': {
+         mime = 'application/x-httpd-php';
+         const printer = json2php.make({ linebreak: '\n', indent: '\t', shortArraySyntax: true });
+         content = printer(args.content);
+         content = `<?php\n$${(args.sqlOptions?.targetTable || args.table).replaceAll('-', '_')} = ${content};`;
          break;
       }
       case 'json':
@@ -54,7 +66,7 @@ export const exportRows = (args: {
 
    const file = new Blob([content], { type: mime });
    const downloadLink = document.createElement('a');
-   downloadLink.download = `${args.table}.${args.type}`;
+   downloadLink.download = `${args.sqlOptions?.targetTable || args.table}.${args.type}`;
    downloadLink.href = window.URL.createObjectURL(file);
    downloadLink.style.display = 'none';
    document.body.appendChild(downloadLink);