+++ src/renderer/untyped.d.ts
@@ -2,6 +2,7 @@
 /* eslint-disable @typescript-eslint/ban-types */
 declare module '@/App.vue';
 declare module 'v-mask';
+declare module 'json2php';
 declare module 'vuedraggable' {// <- to export as default
    const draggableComponent: import('vue').DefineComponent<{
       list: {